feat(feishu): send RUN_ERROR's documented { message, code }, carrying the reason
The operator pointed at the reference the probe had no document to check
against: "COT Message Brief" on the enterprise docs host open.larkoffice.com,
which the public open.feishu.cn documentation does not carry at all. It
confirms the probe and fixes the shape the probe could only guess at.

RUN_FINISHED.status is documented as exactly `done | paused | interrupted`,
which is why a `failed` status rendered 已完成: the client ignores a status it
does not know. `paused` is documented and never produced here — a card is open
or ended, never held — and the omission is noted beside the type.

RUN_ERROR's content is `{ message, code }` and nothing else, so the terminal
drops the threadId/runId the probe sent and puts the failing turn's own reason
in `message`, bounded to 512 bytes because Core bounds a reason at 100_000
characters while one event's content is capped at 4 KiB and a terminal cannot
be chunked. The reason is still printed as ordinary text above the terminal:
whether the client renders this field is unknown, and printing is what makes it
visible. `code` is the fixed category RUN_FAILED, never the reason again.

The terminal intent is now `{ terminal, reason }` — the two fields of the
neutral `turn.ended` — and `detach` takes both at all four call sites; the three
lifecycle paths pass `('interrupted', null)`, because a retired anchor has no
reason and is not a failure.

This needed ~20 lines in a file at 696 of its 700-line cap, so the seam recorded
last commit was taken rather than shaved: feishu-cot-presentation.ts owns what a
card shows for a tool call plus the byte bounding those strings share, and
feishu-cot-events.ts keeps AG-UI event construction and wire budgets. One
direction, no behaviour change, no other file touched.
