+++ common/changes/@excitedjs/feishu-channel/refactor-split-display-from-pushback_2026-09-02-15-00.json
@@ -2,7 +2,7 @@
   "changes": [
     {
       "packageName": "@excitedjs/feishu-channel",
-      "comment": "BREAKING: Review: watch one COT card end-to-end after upgrading, including one input that fails before a runtime accepts it \u2014 that card now reads \u4efb\u52a1\u5931\u8d25 instead of \u5df2\u4e2d\u65ad. No config, state, cache, or path shape changed and no rebuild is needed. The session now subscribes to `teammate.input` and `teammate.activity` instead of the five deleted turn-scoped kinds, and recognizes its own inbound body by comparing `source_id` against the ids this session issued rather than through a two-hop turn correlation. A card's one terminal is a `turn.ended` activity, which prints its reason on the card before ending it. That terminal is now three-valued across two AG-UI events: a `completed` end sends `RUN_FINISHED done`, an `interrupted` end and every lifecycle interrupt send `RUN_FINISHED interrupted`, and a `failed` end sends `RUN_ERROR`, which is what the Feishu client renders as \u4efb\u52a1\u5931\u8d25 \u2014 a submission no runtime accepted used to end the card as \u5df2\u4e2d\u65ad. Open tool calls are keyed on the runtime's call id alone.",
+      "comment": "BREAKING: Review: watch one COT card end-to-end after upgrading, including one input that fails before a runtime accepts it — that card now reads 任务失败 instead of 已中断. No config, state, cache, or path shape changed and no rebuild is needed. The session now subscribes to `teammate.input` and `teammate.activity` instead of the five deleted turn-scoped kinds, and recognizes its own inbound body by comparing `source_id` against the ids this session issued rather than through a two-hop turn correlation. A card's one terminal is a `turn.ended` activity, which prints its reason on the card before ending it. That terminal is now three-valued across two AG-UI events: a `completed` end sends `RUN_FINISHED done`, an `interrupted` end and every lifecycle interrupt send `RUN_FINISHED interrupted`, and a `failed` end sends `RUN_ERROR`, which is what the Feishu client renders as 任务失败 and which carries that end's reason in the event's documented `message` field — a submission no runtime accepted used to end the card as 已中断. Open tool calls are keyed on the runtime's call id alone.",
       "type": "major"
     }
   ],
