+++ packages/channel/feishu-channel/src/feishu-cot-events.ts
@@ -5,25 +5,28 @@
  */
 import { createHash } from 'node:crypto';
 
-import type {
-  RuntimeToolAction,
-  TeammateActivity,
-} from '@excitedjs/dreamux-types';
+import type { TeammateActivity } from '@excitedjs/dreamux-types';
 import type { FeishuCotEventInput } from '@excitedjs/feishu-transport';
 
-/** The one activity member this presentation layer renders. */
-type CotToolCallActivity = Extract<TeammateActivity, { kind: 'tool.call' }>;
+import {
+  escapedBytes,
+  normalizedDetail,
+  toolPresentation,
+  truncateEscaped,
+  truncateUtf8,
+  TOOL_ARGUMENTS_SOFT_MAX_BYTES,
+  TRUNCATION_MARKER,
+  type CotToolCallActivity,
+} from './feishu-cot-presentation.js';
 
 export const FEISHU_COT_EVENT_CONTENT_MAX_BYTES = 4_096;
 export const FEISHU_COT_APPEND_MAX_BYTES = 64 * 1_024;
 
 const TEXT_MESSAGE_EVENT_GROUP_MAX_BYTES = 224 * 1_024;
-const TOOL_NAME_MAX_BYTES = 80;
-const TOOL_ARGUMENTS_SOFT_MAX_BYTES = 512;
 const TOOL_RESULT_SOFT_MAX_BYTES = 1_024;
 const TOOL_RESULT_CONTENT_RESERVE_BYTES = 256;
 const SHORT_TEXT_MAX_BYTES = 120;
-const TRUNCATION_MARKER = '…（已截断）';
+const RUN_ERROR_MESSAGE_MAX_BYTES = 512;
 const COT_EVENT_ENCODING_RESERVE_BYTES = 128;
 const COT_REQUEST_ENCODING_RESERVE_BYTES = 512;
 
@@ -32,11 +35,24 @@ const COT_REQUEST_ENCODING_RESERVE_BYTES = 512;
  * because a card's terminal *is* the end of what it was showing. The lifecycle
  * paths that end a card with no runtime saying so — a retired anchor, a session
  * close — are exactly `interrupted`. Wire spelling is this module's business.
+ *
+ * Feishu documents one more `RUN_FINISHED.status`, `paused`, that nothing here
+ * produces: a card is open or ended, never held. The omission is deliberate.
  */
 export type FeishuCotTerminal = Extract<
   TeammateActivity,
   { kind: 'turn.ended' }
 >['status'];
+
+/**
+ * The terminal a card is waiting to send, held until its outbox drains. Both
+ * fields come straight off the runtime's `turn.ended`; only the failure
+ * terminal has a wire field for the reason, so the other two carry `null`.
+ */
+export interface FeishuCotTerminalIntent {
+  readonly terminal: FeishuCotTerminal;
+  readonly reason: string | null;
+}
 
 export function runStartedEvent(presentationId: string): FeishuCotEventInput {
   return checkedEvent({
@@ -46,23 +62,27 @@ export function runStartedEvent(presentationId: string): FeishuCotEventInput {
 }
 
 /**
- * The one event that ends a card, across two event types: AG-UI puts a failure
- * in `RUN_ERROR` rather than among the values of `RUN_FINISHED.status`, and the
- * Feishu client agrees. Probed live 2026-09-02 — `RUN_FINISHED` renders 已完成
- * for `failed` exactly as it does for a deliberately nonsense status, and only
- * `RUN_ERROR` renders 任务失败. The platform accepted all of them, so only the
- * rendered card is evidence.
+ * The one event that ends a card, across two event types: Feishu documents
+ * `RUN_FINISHED.status` as exactly `done | paused | interrupted` and puts a
+ * failure in its own `RUN_ERROR` event ("COT Message Brief", on the enterprise
+ * docs host `open.larkoffice.com`; the public `open.feishu.cn` docs carry no
+ * `message_cot` reference at all). A live probe agrees: `RUN_FINISHED` with
+ * `failed` renders 已完成, exactly as a deliberately nonsense status does, and
+ * only `RUN_ERROR` renders 任务失败. The platform accepted every one of them,
+ * so only the rendered card is evidence, never the response code.
  *
- * `RUN_ERROR` carries the fixed `message`/`code` the probe verified, not the
- * turn's own reason: that reason is already a text message on this card just
- * above the terminal, so repeating it would show the same text twice.
+ * `RUN_ERROR` sends the documented `{ message, code }` and nothing else: the
+ * probe added `threadId`/`runId` and the platform tolerated them, but they are
+ * not in the reference. `message` is the failing turn's own reason, bounded
+ * here because Core bounds it at 100_000 characters while one event's content
+ * is capped at 4 KiB; `code` is a stable category, not the reason again.
  */
 export function runTerminalEvent(
   presentationId: string,
-  terminal: FeishuCotTerminal,
+  intent: FeishuCotTerminalIntent,
 ): FeishuCotEventInput {
   const run = { threadId: presentationId, runId: presentationId };
-  switch (terminal) {
+  switch (intent.terminal) {
     case 'completed':
       return checkedEvent({
         eventType: 'RUN_FINISHED',
@@ -76,7 +96,12 @@ export function runTerminalEvent(
     case 'failed':
       return checkedEvent({
         eventType: 'RUN_ERROR',
-        content: { ...run, message: '任务失败', code: 'run_failed' },
+        content: {
+          message: intent.reason === null
+            ? '任务失败'
+            : truncateUtf8(intent.reason, RUN_ERROR_MESSAGE_MAX_BYTES),
+          code: 'RUN_FAILED',
+        },
       });
   }
 }
@@ -249,254 +274,6 @@ function opaqueDisplayId(kind: string, source: string): string {
   return `${kind}-${digest}`;
 }
 
-function displayToolName(toolName: string): string {
-  const leaf = toolName
-    .split(/(?:__|[.:/\\])/)
-    .filter((part) => part !== '')
-    .at(-1)
-    ?.trim();
-  return truncateUtf8(
-    leaf === undefined || leaf === '' ? '工具' : leaf,
-    TOOL_NAME_MAX_BYTES,
-  );
-}
-
-type OwnedTool = 'reply' | 'react' | 'list_chat_bots';
-type TeammateTool = 'spawn' | 'send' | 'close' | 'workflow_run';
-
-interface BuiltInToolPresentation {
-  readonly title: string;
-  readonly resultText: string;
-  readonly resultLanguage: 'text' | 'javascript';
-  readonly forceResultCode: boolean;
-}
-
-interface ToolPresentation {
-  readonly toolCallName: string;
-  readonly icon?: 'search' | 'bash' | 'read' | 'write' | 'default';
-  readonly title?: string;
-  readonly ownedTool: OwnedTool | null;
-  readonly builtInTool: BuiltInToolPresentation | null;
-}
-
-const ACTION_TOOL_NAMES: Readonly<Record<RuntimeToolAction, string>> = {
-  read: 'Read',
-  list_files: 'List',
-  search: 'Search',
-  edit: 'Edit',
-  run: 'Bash',
-};
-
-const OWNED_TOOL_PRESENTATION: Readonly<Record<
-  OwnedTool,
-  Pick<ToolPresentation, 'icon' | 'title'>
->> = {
-  reply: { icon: 'write', title: '回复飞书消息' },
-  react: { icon: 'default', title: '点击飞书表情' },
-  list_chat_bots: {
-    icon: 'search',
-    title: '查看群机器人',
-  },
-};
-
-function toolPresentation(
-  event: CotToolCallActivity,
-  channelId: string | undefined,
-): ToolPresentation {
-  const toolCallName = displayToolName(event.tool_name);
-  const ownedTool = ownedFeishuTool(event.tool_name, channelId);
-  if (ownedTool !== null) {
-    return {
-      toolCallName,
-      ...OWNED_TOOL_PRESENTATION[ownedTool],
-      ownedTool,
-      builtInTool: null,
-    };
-  }
-  const builtInTool = builtInToolPresentation(event);
-  if (builtInTool !== null) {
-    return {
-      toolCallName,
-      title: builtInTool.title,
-      ownedTool: null,
-      builtInTool,
-    };
-  }
-  return {
-    toolCallName: event.tool_action === null
-      ? toolCallName
-      : ACTION_TOOL_NAMES[event.tool_action],
-    ownedTool: null,
-    builtInTool: null,
-  };
-}
-
-function ownedFeishuTool(
-  toolName: string,
-  channelId: string | undefined,
-): OwnedTool | null {
-  let server: string | null = null;
-  let leaf: string | null = null;
-  const mcpParts = toolName.split('__');
-  if (mcpParts.length === 3 && mcpParts[0] === 'mcp') {
-    [, server, leaf] = mcpParts;
-  } else {
-    const separator = toolName.lastIndexOf('.');
-    if (separator > 0 && separator < toolName.length - 1) {
-      server = toolName.slice(0, separator);
-      leaf = toolName.slice(separator + 1);
-    }
-  }
-  // Core names a Channel MCP server from the configured channel id, so accept
-  // both the bare id and the namespaced form it is injected under.
-  if (
-    server !== 'feishu' &&
-    (channelId === undefined ||
-      (server !== channelId && server !== `channel-${channelId}`))
-  ) {
-    return null;
-  }
-  return leaf === 'reply' || leaf === 'react' || leaf === 'list_chat_bots'
-    ? leaf
-    : null;
-}
-
-function builtInToolPresentation(
-  event: CotToolCallActivity,
-): BuiltInToolPresentation | null {
-  const tool = teammateTool(event.tool_name);
-  if (
-    tool === null ||
-    event.arguments_truncated ||
-    event.arguments_json === null
-  ) {
-    return null;
-  }
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(event.arguments_json);
-  } catch {
-    return null;
-  }
-  if (typeof parsed !== 'object' || parsed === null || Array.isArray(parsed)) {
-    return null;
-  }
-  const args = parsed as Record<string, unknown>;
-  if (tool === 'spawn') {
-    const namePrefix = nonBlankString(args['name_prefix']);
-    const prompt = nonBlankString(args['prompt']);
-    const intent = stringValue(args['intent']);
-    const agentRuntime = optionalString(args['agent_runtime']);
-    const identity = optionalString(args['identity']);
-    if (
-      namePrefix === null ||
-      prompt === null ||
-      intent === null ||
-      agentRuntime === null ||
-      identity === null
-    ) {
-      return null;
-    }
-    const displayIntent = boundedTitleText(intent);
-    return {
-      title: displayIntent === '' ? '分派成员' : `分派成员 ${displayIntent}`,
-      resultText: [
-        `Agent Runtime：${agentRuntime === undefined || agentRuntime.trim() === ''
-          ? '未指定'
-          : agentRuntime}`,
-        `Identity：${identity === undefined || identity.trim() === ''
-          ? '未指定'
-          : identity}`,
-        `Prompt：${prompt}`,
-      ].join('\n'),
-      resultLanguage: 'text',
-      forceResultCode: false,
-    };
-  }
-  if (tool === 'send') {
-    const name = nonBlankString(args['name']);
-    const prompt = nonBlankString(args['prompt']);
-    if (name === null || prompt === null) return null;
-    return {
-      title: `发送消息 → ${boundedTitleText(name)}`,
-      resultText: `目标：${name}\nPrompt：${prompt}`,
-      resultLanguage: 'text',
-      forceResultCode: false,
-    };
-  }
-  if (tool === 'close') {
-    const name = nonBlankString(args['name']);
-    const note = nonBlankString(args['note']);
-    if (name === null || note === null) return null;
-    return {
-      title: `关闭成员 ${boundedTitleText(name)}`,
-      resultText: note,
-      resultLanguage: 'text',
-      forceResultCode: false,
-    };
-  }
-  const script = nonBlankString(args['script']);
-  const scriptPath = nonBlankString(args['scriptPath']);
-  const resultText = script ?? scriptPath;
-  if (resultText === null) return null;
-  const workflowName = script === null ? null : workflowMetaName(script);
-  return {
-    title: workflowName === null
-      ? 'Workflow'
-      : `Workflow ${boundedTitleText(workflowName)}`,
-    resultText,
-    resultLanguage: script === null ? 'text' : 'javascript',
-    forceResultCode: script !== null,
-  };
-}
-
-function teammateTool(toolName: string): TeammateTool | null {
-  const prefixes = ['mcp__teammate__', 'teammate.'] as const;
-  const prefix = prefixes.find((candidate) => toolName.startsWith(candidate));
-  if (prefix === undefined) return null;
-  const verb = toolName.slice(prefix.length);
-  return verb === 'spawn' || verb === 'send' || verb === 'close' ||
-    verb === 'workflow_run'
-    ? verb
-    : null;
-}
-
-function nonBlankString(value: unknown): string | null {
-  return typeof value === 'string' && value.trim() !== '' ? value : null;
-}
-
-function stringValue(value: unknown): string | null {
-  return typeof value === 'string' ? value : null;
-}
-
-function optionalString(value: unknown): string | null | undefined {
-  return value === undefined || typeof value === 'string' ? value : null;
-}
-
-function boundedTitleText(value: string): string {
-  return normalizedDetail(value, false, TOOL_ARGUMENTS_SOFT_MAX_BYTES)?.trim() ?? '';
-}
-
-function workflowMetaName(script: string): string | null {
-  const match = /\bexport\s+const\s+meta\s*=\s*\{[^}]*?\bname\s*:\s*(['"])([^'"\\\r\n]+)\1/u
-    .exec(script);
-  const name = match?.[2]?.trim();
-  return name === undefined || name === '' ? null : name;
-}
-
-function normalizedDetail(
-  value: string | null,
-  sourceTruncated: boolean,
-  softMaxBytes: number,
-): string | null {
-  if (value === null || value === '') return null;
-  // Preserve provider detail verbatim here; the escaped-byte budget below
-  // remains the authoritative safety boundary for display payload size.
-  const normalized = value;
-  if (normalized === '') return null;
-  if (!sourceTruncated && escapedBytes(normalized) <= softMaxBytes) return normalized;
-  return truncateEscaped(normalized, softMaxBytes, true);
-}
 
 export interface ToolResultParts {
   readonly failed: boolean;
@@ -584,39 +361,6 @@ function shrinkForContentBudget(value: string, overflowBytes: number): string {
   return truncateEscaped(value, target);
 }
 
-function truncateEscaped(
-  value: string,
-  maxBytes: number,
-  forceMarker = false,
-): string {
-  const valueBytes = escapedBytes(value);
-  if (!forceMarker && valueBytes <= maxBytes) return value;
-  const markerBytes = escapedBytes(TRUNCATION_MARKER);
-  if (
-    forceMarker &&
-    value.endsWith(TRUNCATION_MARKER) &&
-    valueBytes <= maxBytes
-  ) {
-    return value;
-  }
-  if (forceMarker && valueBytes + markerBytes <= maxBytes) {
-    return `${value}${TRUNCATION_MARKER}`;
-  }
-  const prefixBudget = Math.max(0, maxBytes - markerBytes);
-  const prefix: string[] = [];
-  let bytes = 0;
-  for (const character of value) {
-    const characterBytes = escapedBytes(character);
-    if (bytes + characterBytes > prefixBudget) break;
-    prefix.push(character);
-    bytes += characterBytes;
-  }
-  return `${prefix.join('')}${TRUNCATION_MARKER}`;
-}
-
-function escapedBytes(value: string): number {
-  return Buffer.byteLength(JSON.stringify(value).slice(1, -1), 'utf8');
-}
 
 function splitForEventContent(
   text: string,
@@ -676,20 +420,6 @@ function boundedTextMessageEventGroup(
   return [start, ...accepted, marker, end];
 }
 
-function truncateUtf8(value: string, maxBytes: number): string {
-  if (Buffer.byteLength(value, 'utf8') <= maxBytes) return value;
-  const markerBytes = Buffer.byteLength(TRUNCATION_MARKER, 'utf8');
-  const prefixBudget = Math.max(0, maxBytes - markerBytes);
-  const prefix: string[] = [];
-  let bytes = 0;
-  for (const character of value) {
-    const characterBytes = Buffer.byteLength(character, 'utf8');
-    if (bytes + characterBytes > prefixBudget) break;
-    prefix.push(character);
-    bytes += characterBytes;
-  }
-  return `${prefix.join('')}${TRUNCATION_MARKER}`;
-}
 
 function jsonBytes(value: unknown): number {
   return Buffer.byteLength(JSON.stringify(value), 'utf8');
